refactor so setup functions are in mod_source_base
