+++ modules/mod_manager.ts
@@ -19,7 +19,6 @@ import { Install, ModList, ModListEntry, ModListLoader } from "./mod_list_loader
 import Main from "../main";
 import ModInstallSource from "./mod_sources/mod_source_base";
 import ElectronLog from "electron-log";
-import CollectionSource from "./mod_sources/collection_source";
 
 var functionMap = new Map();
 
@@ -410,52 +409,7 @@ class ModManager {
         //Save the config changes.
         await config.SaveConfig(Main.config);
         
-        let setup_func: string = "";
-        if(typeof(collectionVersion) != "undefined") {
-            if (!(typeof (this.source_manager.data[FindCollectionNumber(this.source_manager.data, collectionVersion)].setupfunc) == "undefined")) {
-                setup_func = this.source_manager.data[FindCollectionNumber(this.source_manager.data, collectionVersion)].setupfunc;
-            }
-        }
-        
-        switch (setup_func) {
-            case "movecfgs":
-                ElectronLog.log("Executing movecfgs install operation")
-                let filesToMove = ["autoexec.cfg", "scout.cfg", "soldier.cfg", "pyro.cfg", "demoman.cfg", "heavyweapons.cfg",
-                "engineer.cfg", "medic.cfg", "sniper.cfg", "spy.cfg"]
-                //Move cfgs
-                let cfgpath = path.join(Main.config.tf2_directory, "tf", "cfg");
-                let usercfgpath = path.join(cfgpath, "user");
-                ElectronLog.log("cfg path is: " + cfgpath);
-                ElectronLog.log("user cfg path is: \"" + usercfgpath + "\"");
-                if (!(fs.existsSync(usercfgpath))) {
-                    //Create usercfgpath
-                    ElectronLog.log("user cfg path does not exist, creating...");
-                    fs.mkdirSync(usercfgpath, {recursive: true});
-                }
-                //Actually move them
-                fs.readdir(cfgpath, (err, files) => {
-                    if (err) {
-                        ElectronLog.error("there was an error on mod_manager.ts, line 447 when trying to read the cfg directory to move cfg files. Trying to read directory: " + cfgpath + " , the error was:" + err)
-                    }
-                    else {
-                        files.forEach(file => {
-                            ElectronLog.log("Checking if should move file: " + file)
-                            if(filesToMove.indexOf(file) > -1) {
-                                ElectronLog.log("Trying to move file " + path.join(cfgpath, file) + " to " + path.join(usercfgpath, path.basename(file)))
-                                fs.rename(path.join(cfgpath ,file), path.join(usercfgpath, path.basename(file)), (err) => {
-                                    if (err) {
-                                        ElectronLog.error("There was an error trying to move file " + path.join(cfgpath, file) + " to " + path.join(usercfgpath, path.basename(file)))
-                                    }
-                                })
-                            }
-                        })
-                    }
-                })
-                break;
-                        
-            default:
-                break;
-        }
+        this.source_manager.PostInstall(collectionVersion)
 
         this.currentModState = "INSTALLED";
 
@@ -529,44 +483,7 @@ class ModManager {
                 }
                 //Try to execute mod specific operations, like moving tf/user/cfg/class.cfg and tf/user/cfg/autoexec.cfg back to /tf/cfg/class.cfg
                 //and /tf/cfg/autoexec.cfg respectively for Mastercomfig
-                let setup_func: string = "";
-                if (typeof (this.source_manager.data[0]) != "undefined") {
-                    if ((typeof (this.source_manager.data[0].setupfunc)) != "undefined")
-                    setup_func = this.source_manager.data[0].setupfunc;
-                }
-                switch (setup_func) {
-                    case "movecfgs":
-                        //Move them back
-                        ElectronLog.log("Executing movecfgs uninstall operation")
-                        let filesToMove = ["autoexec.cfg", "scout.cfg", "soldier.cfg", "pyro.cfg", "demoman.cfg", "heavyweapons.cfg",
-                        "engineer.cfg", "medic.cfg", "sniper.cfg", "spy.cfg"]
-                        //Move cfgs
-                        let cfgpath = path.join(Main.config.tf2_directory, "tf", "cfg");
-                        let usercfgpath = path.join(cfgpath, "user");
-                        ElectronLog.log("cfg path is: " + cfgpath);
-                        ElectronLog.log("user cfg path is: \"" + usercfgpath + "\"");
-                        //Actually move them
-                        fs.readdir(usercfgpath, (err, files) => {
-                            if (err)
-                                ElectronLog.error("there was an error on mod_manager.ts, line 566 when trying to read the cfg directory to move cfg files. Trying to read directory: " + cfgpath + "the error was:" + err)
-                            else {
-                                files.forEach(file => {
-                                    ElectronLog.log("Checking if should move file: " + file)
-                                    if(filesToMove.indexOf(file) > -1) {
-                                        ElectronLog.log("Trying to move file " + path.join(usercfgpath, file) + " to " + path.join(cfgpath, path.basename(file)))
-                                        fs.rename(path.join(usercfgpath ,file), path.join(cfgpath, path.basename(file)), (err) => {
-                                            if (err) {
-                                                ElectronLog.error("There was an error trying to move file " + path.join(usercfgpath, file) + " to " + path.join(cfgpath, path.basename(file)))
-                                            }
-                                        })
-                                    }
-                                })
-                            }
-                        })
-                        break;                        
-                    default:
-                        break;
-                }
+                this.source_manager.PostUninstall()
                 await Delay(300);        
                 running = false;
                 progressBar.setCompleted();
@@ -1269,6 +1186,11 @@ function IsNotCollection(arg: Install[]): Boolean {
 
 //Find which element in an Install[] has the desiredCollectionString as an argument
 function FindCollectionNumber(Installs: Install[], desiredCollectionString: string) : number {
+    
+    if(typeof(desiredCollectionString) == 'undefined') {
+        return 0
+    }
+
     if(Installs.length < 2) {
         //It has a single item
         return 0
@@ -1289,3 +1211,5 @@ functionMap.set("vpk", WriteFilesToDirectory);
 functionMap.set("rar", async() => { await FatalError("Cannot handle .rar files currently. This should not happen. Exiting..."); });
 functionMap.set("7z", async() => { await FatalError("Cannot handle .7z files currently. This should not happen. Exiting..."); });
 functionMap.set("exe", async() => { await FatalError("Downloaded file was windows executable. This should not happen, exiting. File was not written."); });
+
+export { FindCollectionNumber };
